requirements/prod.txt
aiogram==3.4.1
alembic==1.13.1
cachetools==5.3.3
geopy==2.4.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==5.0.3
requests==2.31.0
sqlalchemy==2.0.28
